dont delete planed sms with recurrent option, fix small bug when hit back on contact selection

+++ app/src/main/java/ch/hesso/l8erproject/l8er/tools/SMSSender.kt
@@ -8,10 +8,14 @@ import android.widget.Toast
 
 val updateDBIntentName = "DB-UPDATE"
 
-fun sendSMS(context: Context, smsId: Int, number: String, text: String) {
+fun sendSMS(context: Context, smsId: Int, number: String, text: String, mustDelete: Boolean) {
     SmsManager.getDefault().sendTextMessage(number, null, text, null, null)
     val smsDBHelper = SMSDBHelper(context)
-    smsDBHelper.deleteSMS(smsId)
+
+    if (mustDelete) {
+        smsDBHelper.deleteSMS(smsId)
+    }
+    
     Toast.makeText(context, "sms sent.", Toast.LENGTH_SHORT).show()
 
     val intent: Intent = Intent(updateDBIntentName)
